03_document_chatbot/requirements.txt
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
langchain>=0.0.280,<1.0.0
chromadb>=0.3.29,<1.0.0
tiktoken>=0.4.0,<1.0.0
